Correct "Handicap" column explanation and visual
After going on Kaggle, it turns out that numbers 0 to 4 actually refer
to the number of disabilities a person has, rather than the severity, So
correction was needed to reflect that. Along the way I also discovered
that the numbers were out of order with the visual, giving a false
bar chart. The conclusion was a little changed after correction.